added stt will update help text later after testing
Obtain API key from https://console.bluemix.net/docs/services/speech-to-text/getting-started.html

Sign Up / Log In for the service.

Open this URL after logging in: https://cloud.ibm.com/catalog/services/speech-to-text

Click on the Create button. It is at the right side of the screen, when opened in Desktop mode.

In the Left Side Nav Menu, click on Manage button.

Now, in the main screen, you will be able to see API Key and URL.

There will be a copy button to the right side of the text field.

Add these to the corresponding Heroku Environment Variable keys.

IBM_WATSON_CRED_URL and IBM_WATSON_CRED_PASSWORD